Refactor status condition application to use PlayedCard method
Extracts status condition logic into a PlayedCard::apply_status_condition
method to improve encapsulation and code reuse. The Arceus ex immunity
check is now handled by the PlayedCard itself rather than being
duplicated in multiple places.

Changes:
- Added apply_status_condition method to PlayedCard
- Made StatusCondition implement Copy for easier use in closures
- Updated build_status_effect to delegate to PlayedCard method
- Updated mega_burning_attack to use the new method directly
- Removed duplicate Arceus immunity logic from build_status_effect

Benefits:
- Single source of truth for status application rules
- Arceus immunity is consistently enforced
- Cleaner, more maintainable code
