+++ src/models/card.rs
@@ -217,7 +217,7 @@ impl Card {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 pub enum StatusCondition {
     Poisoned,
     Paralyzed,
